requirements.txt
pandas
numpy
seaborn
matplotlib
google-cloud-bigquery
plotly-express
streamlit